Rename files ros_controllers_widget.* -> controllers_widget.*
