+++ moveit_setup_assistant/src/widgets/controllers_widget.h
